chore rendering after checking auth
