+++ app/products/[id]/page-components/ProductContent.tsx
@@ -5,7 +5,13 @@ import { ProductDetailResponse } from '@/hooks/query/useProductDetail';
 import Image from 'next/image';
 
 // 상품 상세 내용 컴포넌트
-const ProductContent = ({ product }: { product: ProductDetailResponse }) => {
+const ProductContent = ({
+  product,
+  isOwner,
+}: {
+  product: ProductDetailResponse;
+  isOwner: boolean;
+}) => {
   const { content, product_tags, count, thumbnail_url } = product;
 
   const handleShare = () => {
